Crate sc_executor_wasmi

This crate provides an implementation of WasmModule that is baked by wasmi.

Structs

WasmiInstance
Wasmi instance wrapper along with the state snapshot.

WasmiRuntime
A runtime along with initial copy of data segments.

Functions

create_runtime
Create a new WasmiRuntime given the code. This function loads the module and stores it in the instance.